+++ app/Http/Controllers/MapController.php
@@ -4,18 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-//Siterモデルの読み込み
-use App\Models\Site;
 
 class MapController extends Controller
 {
 
     public function view()
     {
-        $sites = Site::all();
-        return view('map.view', [
-            'sites' => $sites
-        ]);
+        return view('map.view');
     }
 
     public function post()
